page-based/data-collection/requirements.txt
adblockparser==0.7
argh==0.26.2
asn1crypto==0.24.0
beautifulsoup4==4.6.3
blinker==1.4
brotlipy==0.7.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.2
cloudpickle==0.6.1
coloredlogs==10.0
construct==2.5.3
cryptography==2.3.1
cssutils==1.0.2
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
EasyProcess==0.2.3
h11==0.7.0
h2==3.0.1
hpack==3.0.0
html2text==2015.11.4
humanfriendly==4.17
hyperframe==5.1.0
idna==2.7
jsbeautifier==1.6.14
kaitaistruct==0.8
kiwisolver==1.0.1
ldap3==2.5.1
mitmproxy==4.0.4
networkx==2.2
numpy==1.15.4
packaging==18.0
passlib==1.7.1
pathtools==0.1.2
Pillow==5.3.0
pyasn1==0.4.4
pycparser==2.19
pymongo==3.7.2
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
python-dateutil==2.7.5
python-magic==0.4.15
PyVirtualDisplay==0.2.1
PyWavelets==1.0.1
PyYAML==3.13
requests==2.20.1
ruamel.yaml==0.13.14
scikit-image==0.14.1
scipy==1.1.0
six==1.10.0
sortedcontainers==1.5.10
toolz==0.9.0
tornado==4.4.3
urllib3==1.24.1
urwid==2.0.1
watchdog==0.8.3
wsproto==0.11.0
